# api/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  description: ESIR2 Projet Simba - Smart Doodle
  version: 1.0.0
  title: Swagger Projet Simba
...
paths:
  /api/users:
    post:
      tags:
        - Users
      summary: Create a new user
      operationId: createUser
      x-@id: evolvable:createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreationRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          links:
            commentPoll:
              operationId: commentPoll
              x-@relation: evolvable:rel/nextComment
              parameters:
                userId: $response.body#/id
            vote:
              operationId: vote
              x-@relation: evolvable:rel/nextVote
              x-affiliation: parent
              parameters:
                userId: $response.body#/id
  /api/polls:
    post:
      tags:
        - Polls
      summary: Create a new poll
      operationId: createPoll
      x-@id: evolvable:createPoll
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PollCreationRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Poll'
          links:
            comment:
              operationId: createUser
              x-@relation: evolvable:rel/comment
              x-affiliation: parent
            vote:
              operationId: createUser
              x-@relation: evolvable:rel/vote
              x-affiliation: parent
            addChoice:
              operationId: createChoices
              x-@relation: evolvable:rel/addChoice
              x-affiliation: parent
              parameters:
                slug: $response.body#/slug
            deleteChoice:
              operationId: deleteChoiceFromPoll
              x-@relation: evolvable:rel/deleteChoice
              x-affiliation: parent
              parameters:
                slug: $response.body#/slug
            updateChoice:
              operationId: updateChoice
              x-@relation: evolvable:rel/updateChoice
              x-affiliation: parent
              parameters:
                slug: $response.body#/slug
  /api/polls/{slug}:
    get:
      tags:
        - Polls
      summary: Get a specific poll by id
      operationId: retrievePoll
      x-@id:
        - evolvable:getPoll
        - evolvable:getPollBySlug
        - evolvable:getPollWithoutAdminSlug
      parameters:
        - in: path
          name: slug
          schema:
            type: string
          required: true
        - in: query
          name: token
          schema:
            type: string
          required: false
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Poll'
          links:
            comment:
              operationId: createUser
              x-@relation: evolvable:rel/comment
              x-affiliation: parent
            vote:
              operationId: createUser
              x-@relation: evolvable:rel/vote
              x-affiliation: parent
            addChoice:
              operationId: createChoices
              x-@relation: evolvable:rel/addChoice
              x-affiliation: parent
              parameters:
                slug: $response.body#/slug
            deleteChoice:
              operationId: deleteChoiceFromPoll
              x-@relation: evolvable:rel/deleteChoice
              x-affiliation: parent
              parameters:
                slug: $response.body#/slug
            updateChoice:
              operationId: updateChoice
              x-@relation: evolvable:rel/updateChoice
              x-affiliation: parent
              parameters:
                slug: $response.body#/slug
        '401':
          description: Unauthorized
        '404':
          description: Not Found


    put:
      tags:
        - Polls
      summary: Update a specific poll
      operationId: updatePoll
      x-@id: evolvable:updatePoll
      parameters:
        - in: path
          name: slug
          schema:
            type: string
          required: true
        - in: query
          name: token
          schema:
            type: string
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PollUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Poll'
          links:
            comment:
              operationId: createUser
              x-@relation: evolvable:rel/comment
              x-affiliation: parent
            vote:
              operationId: createUser
              x-@relation: evolvable:rel/vote
              x-affiliation: parent
            addChoice:
              operationId: createChoices
              x-@relation: evolvable:rel/addChoice
              x-affiliation: parent
              parameters:
                slug: $response.body#/slug
            deleteChoice:
              operationId: deleteChoiceFromPoll
              x-@relation: evolvable:rel/deleteChoice
              x-affiliation: parent
              parameters:
                slug: $response.body#/slug
            updateChoice:
              operationId: updateChoice
              x-@relation: evolvable:rel/updateChoice
              x-affiliation: parent
              parameters:
                slug: $response.body#/slug
        '401':
          description: Unauthorized
        '404':
          description: Not Found
...
components:
  schemas:
    User:
      type: object
      x-@type: evolvable:User
      required:
        - id
        - username
        - userChoices
        - userPolls
        - userMealPreferences
      properties:
        id:
          type: number
          format: long
        username:
          type: string
        userChoices:
          type: array
          default: []
          items:
            $ref: '#/components/schemas/Choice'
        userPolls:
          type: array
          default: []
          items:
            $ref: '#/components/schemas/Poll'
        userMealPreferences:
          type: array
          default: []
          items:
            $ref: '#/components/schemas/Choice'
    UserCreationRequest:
      type: object
      required:
        - username
      properties:
        username:
          type: string
    Choice:
      type: object
      x-@type: evolvable:Choice
      required:
        - id
        - startDate
        - endDate
        - users
      properties:
        id:
          x-@id:
            - evolvable:id
            - evolvable:choiceId
          x-@type:
            - evolvable:Id
            - evolvable:ChoiceId
          type: number
          format: long
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        users:
          $ref: '#/components/schemas/Users'
      links:
        delete:
          operationId: deleteChoiceFromPoll
          x-@relation: evolvable:rel/delete
          x-affiliation: parent
          parameters:
            slug: $request.path.slug
            choices:
              - $response.body/#id
        update:
          operationId: updateChoice
          x-@relation: evolvable:rel/update
          x-affiliation: parent
          parameters:
            slug: $request.path.slug
            choiceId: $response.body/#id
...
    Choices:
      type: array
      x-@type: evolvable:Choices
      default: []
      items:
        $ref: '#/components/schemas/Choice'
    Comments:
      type: array
      x-@type: evolvable:Comments
      default: []
      items:
        $ref: '#/components/schemas/Comment'
    MealPreferences:
      type: array
      x-@type: evolvable:MealPreferences
      default: []
      items:
        $ref: '#/components/schemas/MealPreference'
    Users:
      type: array
      x-@type: evolvable:Users
      default: []
      items:
        $ref: '#/components/schemas/User'
    Poll:
      type: object
      x-@type: evolvable:Poll
      required:
        - id
        - title
        - location
        - has_meal
        - slug
        - createdAt
        - updatedAt
        - pollChoices
      properties:
        id:
          type: number
          format: long
        title:
          type: string
        location:
          type: string
        description:
          type: string
        has_meal:
          type: boolean
        slug:
          type: string
        slugAdmin:
          type: string
        tlkURL:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        pollChoices:
          $ref: '#/components/schemas/Choices'
        pollComments:
          $ref: '#/components/schemas/Comments'
        pollMealPreferences:
          $ref: '#/components/schemas/MealPreferences'
        padURL:
          type: string
    PollCreationRequest:
      type: object
      required:
        - title
        - description
        - location
        - has_meal
        - pollChoices
      properties:
        title:
          type: string
        description:
          type: string
        location:
          type: string
        has_meal:
          type: boolean
        pollChoices:
          $ref: '#/components/schemas/Choices'
    PollUpdateRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        location:
          type: string
        has_meal:
          type: boolean
    Comment:
      type: object
      x-@type: evolvable:Comment
      required:
        - id
        - content
        - user
      properties:
        id:
          type: number
          format: long
        content:
          type: string
        user:
          $ref: '#/components/schemas/User'
...
    MealPreference:
      type: object
      required:
        - id
        - content
        - user
      properties:
        id:
          type: number
          format: long
        content:
          type: string
        user:
          $ref: '#/components/schemas/User'


# json-LD context
...